claw-miner/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawminer"
version = "0.1.0"
description = "ClawNetwork Agent Mining CLI — earn CLAW by contributing to the network"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "pynacl>=1.5",
    "tomli>=2.0",
    "tomli-w>=1.0",
    "rich>=13.0",
    "blake3>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "respx>=0.20",
]

[project.scripts]
claw-miner = "clawminer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["clawminer"]
